+++ sveltekit/src/routes/import/+page.svelte
@@ -1,7 +1,5 @@
 <script lang="ts">
-	import { onMount } from 'svelte';
 	import { page } from '$app/stores';
-	import type { PageData } from './$types';
 
 	import ScrollView from '$lib/components/ScrollView.svelte';
 	import Section from '$lib/components/Section.svelte';
@@ -15,36 +13,15 @@
 	import FormInput from '$lib/components/FormInput.svelte';
 	import statusBarStore from '$lib/stores/statusBarStore';
 	import FormFooter from '$lib/components/FormFooter.svelte';
-	import FormNotice from '$lib/components/FormNotice.svelte';
-	import FormNoticeNotice from '$lib/components/FormNoticeNotice.svelte';
-	import FormNoticeP from '$lib/components/FormNoticeP.svelte';
-	import FormSelect from '$lib/components/FormSelect.svelte';
 	import { api } from '$lib/helpers/misc';
 	import { CardAppearance } from '$lib/components/Card';
-	import { SyncSettings, Appearance, EventFrequency } from '$lib/helpers/constants';
-	import type { ImportSummary, ImportSync } from '$lib/helpers/import';
+	import { Appearance } from '$lib/helpers/constants';
+	import type { ImportSummary } from '$lib/helpers/import';
 
 	const title = 'Import CanutinFile';
 
-	// Set the loading state
-	const setLoadingStatus = (message: string) => {
-		isImportLoading = true;
-		$statusBarStore = {
-			message,
-			appearance: Appearance.ACTIVE
-		};
-	};
-
-	const setResultStatus = (message: string, appearance: Appearance) => {
-		isImportLoading = false;
-		$statusBarStore = {
-			message,
-			appearance
-		};
-	};
-
 	// File form
-	let isImportLoading: boolean = false;
+	let isLoading: boolean = false;
 	let noFileError: string | undefined = undefined;
 	let importSummary: ImportSummary | undefined = undefined;
 
@@ -57,7 +34,12 @@
 			return;
 		}
 
-		setLoadingStatus('Processing import...');
+		// Set the loading state
+		isLoading = true;
+		$statusBarStore = {
+			message: 'Processing import...',
+			appearance: Appearance.ACTIVE
+		};
 
 		const reader = new FileReader();
 		reader.onload = async (event: ProgressEvent<FileReader>) => {
@@ -65,91 +47,15 @@
 			importSummary = await api({ endpoint: 'import', method: 'POST', payload: canutinFile });
 
 			// Update the status bar
-			setResultStatus(
-				importSummary?.error ? importSummary.error : 'Import was successful',
-				importSummary?.error ? Appearance.NEGATIVE : Appearance.POSITIVE
-			);
+			isLoading = false;
+			$statusBarStore = {
+				message: importSummary?.error ? importSummary.error : 'Import was successful',
+				appearance: importSummary?.error ? Appearance.NEGATIVE : Appearance.POSITIVE
+			};
 		};
 
 		reader.readAsText(chosenFile);
 	};
-
-	// Sync form
-	export let data: PageData;
-	$: ({ settings } = data);
-
-	$: isSyncEnabled = false;
-	$: canutinFileUrlValue = '';
-	$: frequencyValue = 0;
-	$: cookieValue = '';
-	$: jwtValue = '';
-
-	const handleSyncForm = async (event: any) => {
-		const canutinFileUrl = event.target.canutinFileUrl?.value;
-		const frequency = event.target.frequency?.value;
-		const cookie = event.target.cookie?.value;
-		const jwt = event.target.jwt?.value;
-
-		if (!canutinFileUrl) return;
-
-		const payload: ImportSync = {
-			canutinFileUrl,
-			frequency,
-			cookie,
-			jwt
-		};
-
-		!isSyncEnabled && setLoadingStatus('Checking the CanutinFile URL...');
-
-		const response = await api({
-			endpoint: 'sync',
-			method: 'POST',
-			payload: payload
-		});
-
-		// Update the status bar
-		setResultStatus(
-			response?.error
-				? response?.error
-				: response?.warning
-				? response.warning
-				: 'Sync was enabled succesfully',
-			response?.error
-				? Appearance.NEGATIVE
-				: response?.warning
-				? Appearance.WARNING
-				: Appearance.POSITIVE
-		);
-
-		isSyncEnabled = response.isSyncEnabled;
-	};
-
-	const frequencyOptions = Object.values(EventFrequency).map((value) => ({
-		label: value
-	}));
-
-	onMount(async () => {
-		settings.forEach((setting) => {
-			if (!setting) return;
-			switch (setting.name) {
-				case SyncSettings.SYNC_ENABLED:
-					isSyncEnabled = setting.value === '1' ? true : false;
-					break;
-				case SyncSettings.SYNC_URL:
-					canutinFileUrlValue = setting.value;
-					break;
-				case SyncSettings.SYNC_FREQUENCY:
-					frequencyValue = parseInt(setting.value);
-					break;
-				case SyncSettings.SYNC_COOKIE:
-					cookieValue = setting.value;
-					break;
-				case SyncSettings.SYNC_JWT:
-					jwtValue = setting.value;
-					break;
-			}
-		});
-	});
 </script>
 
 <svelte:head>
@@ -157,70 +63,6 @@
 </svelte:head>
 
 <ScrollView {title}>
-	<Section title="Sync settings">
-		<div slot="CONTENT" class="import">
-			<Form on:submit={handleSyncForm}>
-				<FormFieldset>
-					<FormField name="status" label="Sync status">
-						<FormNotice>
-							<FormNoticeNotice
-								appearance={isSyncEnabled ? Appearance.POSITIVE : Appearance.WARNING}
-							>
-								<FormNoticeP>
-									<strong>Sync is {isSyncEnabled ? 'enabled' : 'disabled'}</strong>
-								</FormNoticeP>
-								{#if !isSyncEnabled}
-									<FormNoticeP>
-										Enable syncing by providing a URL that can be fetched as a CanutinFile JSON
-										payload. On every sync the vault will be updated with any new data found,
-										duplicates will be ignored.
-									</FormNoticeP>
-								{/if}
-							</FormNoticeNotice>
-						</FormNotice>
-					</FormField>
-				</FormFieldset>
-				<FormFieldset>
-					<FormField name="canutinFileUrl" label="CanutinFile URL">
-						<FormInput
-							name="canutinFileUrl"
-							placeholder="https://example.com/my-scraper/canutinFile.json"
-							bind:value={canutinFileUrlValue}
-						/>
-					</FormField>
-					<FormField name="cookie" label="Cookie" optional={true}>
-						<FormInput
-							required={false}
-							name="cookie"
-							type="password"
-							placeholder="accessToken=1234abc; userId=1234; Path=/; HttpOnly;"
-							value={cookieValue}
-						/>
-					</FormField>
-					<FormField name="jwt" label="JSON Web Token" optional={true}>
-						<FormInput
-							type="password"
-							required={false}
-							name="jwt"
-							value={jwtValue}
-							placeholder="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
-						/>
-					</FormField>
-				</FormFieldset>
-				<FormFieldset>
-					<FormField name="frequency" label="Auto-sync">
-						<FormSelect name="frequency" options={frequencyOptions} bind:value={frequencyValue} />
-					</FormField>
-				</FormFieldset>
-				<FormFooter>
-					<Button appearance={Appearance.ACTIVE} disabled={!canutinFileUrlValue}
-						>{isSyncEnabled ? 'Update' : 'Enable'}</Button
-					>
-				</FormFooter>
-			</Form>
-		</div>
-	</Section>
-
 	<Section title="From API">
 		<div slot="CONTENT">
 			<Notice>
@@ -246,7 +88,7 @@
 		</div>
 	</Section>
 
-	{#if !isImportLoading && !error && importSummary}
+	{#if !isLoading && !error && importSummary}
 		<div class="importStatus">
 			<div class="importStatus__model">
 				<Section title="Accounts">
